+++ src/main/java/ru/innopolis/stc9/services/MarkService.java
@@ -8,11 +8,15 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
+/**
+ * Сервис оценок
+ */
 public class MarkService implements ObjectService {
     ObjectsDAO objectsDAO = new MarksDAO();
 
     public List<String> isCorrectData(Map<String, String[]> incParam) {
         List<String> result = new ArrayList<>();
+        if (incParam == null || incParam.isEmpty()) return result;
         int value = 0;
         try {
             if (incParam.get("value") != null) value = Integer.parseInt(incParam.get("value")[0]);
@@ -39,18 +43,23 @@ public class MarkService implements ObjectService {
     }
 
     public List<DBObject> getObjects(Map<String, String[]> incParam, String stopWord) {
+        if ((incParam == null) || incParam.isEmpty() || (stopWord == null) || !incParam.containsKey(stopWord))
+            return new ArrayList<>();
         return objectsDAO.getAllByParam(incParam, stopWord);
     }
 
     public void addObject(Map<String, String[]> incParam) {
+        if ((incParam == null) || incParam.isEmpty()) return;
         objectsDAO.addObject(incParam);
     }
 
     public void delObject(Map<String, String[]> incParam) {
+        if ((incParam == null) || incParam.isEmpty()) return;
         objectsDAO.deleteObjectById(Integer.parseInt(incParam.get("id")[0]));
     }
 
     public void updateObject(Map<String, String[]> incParam) {
+        if ((incParam == null) || incParam.isEmpty()) return;
         objectsDAO.updateObject(incParam);
     }
 }
