+++ src/main/java/ru/innopolis/stc9/services/StudentService.java
@@ -8,11 +8,15 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
+/**
+ * Сервис студентов
+ */
 public class StudentService implements ObjectService {
     private static ObjectsDAO objectsDAO = new StudentsDAO();
 
     public List<String> isCorrectData(Map<String, String[]> incParam) {
         List<String> result = new ArrayList<>();
+        if (incParam == null || incParam.isEmpty()) return result;
         if ((incParam.get("firstName") != null) && incParam.get("firstName")[0].equals(""))
             result.add("Invalid first name");
         if ((incParam.get("secondName") != null) && incParam.get("secondName")[0].equals(""))
@@ -33,18 +37,23 @@ public class StudentService implements ObjectService {
     }
 
     public List<DBObject> getObjects(Map<String, String[]> incParam, String stopWord) {
+        if ((incParam == null) || incParam.isEmpty() || (stopWord == null) || !incParam.containsKey(stopWord))
+            return new ArrayList<>();
         return objectsDAO.getAllByParam(incParam, stopWord);
     }
 
     public void addObject(Map<String, String[]> incParam) {
+        if ((incParam == null) || incParam.isEmpty()) return;
         objectsDAO.addObject(incParam);
     }
 
     public void delObject(Map<String, String[]> incParam) {
+        if ((incParam == null) || incParam.isEmpty()) return;
         objectsDAO.deleteObjectById(Integer.parseInt(incParam.get("id")[0]));
     }
 
     public void updateObject(Map<String, String[]> incParam) {
+        if ((incParam == null) || incParam.isEmpty()) return;
         objectsDAO.updateObject(incParam);
     }
 }
